Move InputToken.svelte to ui-components
